+++ ServersideQoL.AutoProcess/CHANGELOG.md
@@ -1,2 +1,5 @@
+### v2.0.11
+- Move `FeedFromContainersMaxRange` config option from ContainerSigns to AutoProcess
+
 ### v2.0.0
 - Initial release