+++ ServersideQoL.AutoStore/CHANGELOG.md
@@ -1,3 +1,6 @@
+### v2.0.11
+- Move `AutoPickupMaxRange` config option from ContainerSigns to AutoStore
+
 ### v2.0.8
 - Bugfixes